Factored out OneSignTextWatcher
