+++ collect_app/src/main/java/org/odk/collect/android/projects/AddProjectDialog.kt
@@ -2,8 +2,6 @@ package org.odk.collect.android.projects
 
 import android.content.Context
 import android.os.Bundle
-import android.text.Editable
-import android.text.TextWatcher
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -11,6 +9,7 @@ import androidx.appcompat.widget.Toolbar
 import org.odk.collect.android.R
 import org.odk.collect.android.databinding.AddProjectDialogLayoutBinding
 import org.odk.collect.android.injection.DaggerUtils
+import org.odk.collect.androidshared.OneSignTextWatcher
 import org.odk.collect.material.MaterialFullScreenDialogFragment
 import org.odk.collect.projects.Project
 import org.odk.collect.projects.ProjectsRepository
@@ -43,25 +42,7 @@ class AddProjectDialog : MaterialFullScreenDialogFragment() {
         super.onViewCreated(view, savedInstanceState)
         setUpToolbar()
 
-        lateinit var oldTextString: String
-        binding.projectIconInputText.addTextChangedListener(object : TextWatcher {
-            override fun beforeTextChanged(charSequence: CharSequence, i: Int, i1: Int, i2: Int) {
-                oldTextString = charSequence.toString()
-            }
-
-            override fun onTextChanged(charSequence: CharSequence, i: Int, i1: Int, i2: Int) {}
-
-            override fun afterTextChanged(editable: Editable) {
-                var newTextString = editable.toString()
-                if (oldTextString != newTextString) {
-                    if (Character.codePointCount(newTextString, 0, newTextString.length) > 1) {
-                        newTextString = oldTextString
-                    }
-                    binding.projectIconInputText.setText(newTextString)
-                    binding.projectIconInputText.setSelection(newTextString.length)
-                }
-            }
-        })
+        binding.projectIconInputText.addTextChangedListener(OneSignTextWatcher(binding.projectIconInputText))
 
         binding.cancelButton.setOnClickListener {
             dismiss()
